+++ src/app.js
@@ -27,6 +27,7 @@ const transactionsRouter = require("./routes/transactions");
 const usersRouter = require("./routes/users");
 const verificationsRouter = require("./routes/verifications");
 const analyticsRouter = require("./routes/analytics");
+const statementRouter = require("./routes/statements");
 
 app.use("/api/v1/auth", auth);
 app.use("/api/v1/auth/google", googleAuth);
@@ -38,9 +39,10 @@ app.use("/api/v1/transactions", transactionsRouter);
 app.use("/api/v1/users", usersRouter);
 app.use("/api/v1/verifications", verificationsRouter);
 app.use("/api/v1/analytics", analyticsRouter);
+// app.use("/api/v1/statement", statementRouter);
 // error handler
 app.use(errorHandler);
-
+/*
 sequelize
   .sync({ alter: true })
   .then(() => {
@@ -49,5 +51,5 @@ sequelize
   .catch((err) => {
     console.log("Error syncing models", err);
   });
-
+*/
 module.exports = app;
