+++ src/controllers/googleAuthController.js
@@ -1,4 +1,4 @@
-const authService = require("../services/auth");
+const authService = require("../services/authService");
 const userService = require("../services/user");
 
 const {
