+++ src/controllers/statementController.js
@@ -1,0 +1,5 @@
+const userService = require('../services/userService');
+
+const downloadStatement = async (req, res, next) => {
+    const { transactionType, startDate, endDate, email } = req.body;
+
