+++ src/services/analyticsFormatterService.js
